// app/shared/enums.ts
 export enum StatusOfContract {
  All = 0,
  Pendiente = 1,
  Activo = 2,
  Terminado = 3,
  Renovado = 4,
  Cancelado = 5
}

export enum StatusOfContractCancel {
  Solicitado = 1,
  Aprobado = 2,
  Cerrado = 3,
  Cancelado = 4
}

export enum PayType {
  Pago = 1,
  Reinversion = 2
}

export enum StatusOfPay {
  Pendiente = 1,
  Terminado = 2,
  Cancelado = 3,
}